`bugfix`: stop servers before writing binary, and correct compare logic
This commit adds a step to stop language servers before writing the new
binary to the language server file. This fixes an issue where on linux
machines, the download fails due to the binary being in use.

This also corrects the logic for when to download the latest release
of the language server.