Add reformat functionality for jsonnet/libsonnet files.
This was a bit of a complicated fix, as lsp4intellij uses an outdated method
to implement formatting (they override the entire `Reformat Code` action.
Instead, we add a `formattingservice` with an ExternalFormatter implementation.

Additionally, the way lsp4intellij reformats code had to be adjusted. From inline
comment:

When the language server returns edits, often the edit entries will have the same
range/span (i.e. same line). However, the order of the edits is still correct for insertion.
But, the edits in lsp4intellij are applied bottom to top so the result ordering
is incorrect. To fix this, custom reformat logic is added to reverse the order
of entries with the same line/range/span which achieves the correct order of inserts.