chore: update plugin for Intellij 2023.3
We were updating the plugin to be able to run with Intellij 2023.3
versions. We had to include a few system upgrades to facilitate bumping
the maximum build version.

1. Upgrade the Gradle version to 8.5 to overcome the error:
```
Unsupported class file major version 63
```

2. Upgrade the Gradle Kotlin and IntelliJ plugins to fix the error:
```
The binary version of its metadata is 1.9.0, and the expected version is 1.7.1.
```

I've followed the instructions in
https://intellij-support.jetbrains.com/hc/en-us/community/posts/15634389368722-After-updating-to-intellij-2023-3-my-plugin-is-incompatible-and-does-not-build-anymore
and https://stackoverflow.com/a/75180866/1132127 to solve the problem.